refactor: restructure and move code logic from FilterTab into ListScroll
